google-cloud-functions/gke-empty-nodes/requirements.txt
kubernetes
google-cloud-container
google-cloud-datastore
